+++ tests/test_aio.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from aiohttp_retry import RetryClient
 from yarl import URL
 from yt_dlp_utils.aio import AsyncYoutubeDL, get_configured_yt_dlp, setup_session
 from yt_dlp_utils.constants import SHARED_HEADERS
@@ -24,6 +25,7 @@ async def test_setup_session_no_domains(mocker: MockerFixture) -> None:
     session = await setup_session(browser='chrome', profile='default')
 
     try:
+        assert isinstance(session, aiohttp.ClientSession)
         assert session.headers['user-agent'] == SHARED_HEADERS['user-agent']
         assert 'cookie1' in session.cookie_jar.filter_cookies(URL('https://example.com'))
     finally:
@@ -41,6 +43,7 @@ async def test_setup_session_with_domains(mocker: MockerFixture) -> None:
     session = await setup_session(browser='chrome', profile='default', domains=['example.com'])
 
     try:
+        assert isinstance(session, aiohttp.ClientSession)
         mock_jar.get_cookies_for_url.assert_called_once_with('https://example.com')
         assert session.headers['user-agent'] == SHARED_HEADERS['user-agent']
     finally:
@@ -60,6 +63,7 @@ async def test_setup_session_no_domains_non_string_cookie(mocker: MockerFixture)
     session = await setup_session(browser='chrome', profile='default')
 
     try:
+        assert isinstance(session, aiohttp.ClientSession)
         cookies = session.cookie_jar.filter_cookies(URL('https://example.com'))
         assert 'cookie1' in cookies
         assert 'cookie2' not in cookies
@@ -80,6 +84,7 @@ async def test_setup_session_with_domains_non_string_cookie(mocker: MockerFixtur
     session = await setup_session(browser='chrome', profile='default', domains=['example.com'])
 
     try:
+        assert isinstance(session, aiohttp.ClientSession)
         cookies = session.cookie_jar.filter_cookies(URL('https://example.com'))
         assert 'cookie1' in cookies
         assert 'cookie2' not in cookies
@@ -99,6 +104,7 @@ async def test_setup_session_with_custom_headers(mocker: MockerFixture) -> None:
                                   add_headers={'Another-Header': 'AnotherValue'})
 
     try:
+        assert isinstance(session, aiohttp.ClientSession)
         assert session.headers['Custom-Header'] == 'CustomValue'
         assert session.headers['Another-Header'] == 'AnotherValue'
     finally:
@@ -114,6 +120,7 @@ async def test_setup_session_with_existing_session(mocker: MockerFixture) -> Non
     existing_session = aiohttp.ClientSession()
     try:
         session = await setup_session(browser='chrome', profile='default', session=existing_session)
+        assert isinstance(session, aiohttp.ClientSession)
         assert session is existing_session
         assert session.headers['user-agent'] == SHARED_HEADERS['user-agent']
     finally:
@@ -207,3 +214,21 @@ def test_async_youtube_dl_ydl_property(mocker: MockerFixture) -> None:
     async_ydl = AsyncYoutubeDL(mock_ydl)
 
     assert async_ydl.ydl is mock_ydl
+
+
+@pytest.mark.asyncio
+async def test_setup_session_with_retry(mocker: MockerFixture) -> None:
+    mock_jar = mocker.Mock(spec=yt_dlp.cookies.YoutubeDLCookieJar)
+    mock_jar.__iter__ = mocker.Mock(return_value=iter([]))
+    mocker.patch('yt_dlp_utils.aio.extract_cookies_from_browser', return_value=mock_jar)
+
+    client = await setup_session(browser='chrome',
+                                 profile='default',
+                                 setup_retry=True,
+                                 backoff_factor=0.5,
+                                 status_forcelist=[500, 502, 503])
+
+    try:
+        assert isinstance(client, RetryClient)
+    finally:
+        await client.close()